move datasets from lir/data/datasets to lir/data
