+++ docs/lrsystem-assessment.rst
@@ -129,7 +129,7 @@ distribution. It visualizes as follows.
 
     from lir.algorithms.logistic_regression import LogitCalibrator
     from lir.data_strategies import TrainTestSplit
-    from lir.data.datasets.synthesized_normal_binary import SynthesizedNormalData, SynthesizedNormalBinaryData
+    from lir.datasets.synthesized_normal_binary import SynthesizedNormalData, SynthesizedNormalBinaryData
 
     # set the parameters for H1 data and H2 data
     h1_data = SynthesizedNormalData(mean=1, std=1, size=1000)
