+++ tests/test_data_strategies.py
@@ -1,11 +1,11 @@
 import numpy as np
 
-from lir.data.datasets.synthesized_normal_multiclass import (
+from lir.data.models import FeatureData
+from lir.data_strategies import PairsTrainTestSplit, PredefinedTrainTestSplit, SourcesTrainTestSplit
+from lir.datasets.synthesized_normal_multiclass import (
     SynthesizedDimension,
     SynthesizedNormalMulticlassData,
 )
-from lir.data.models import FeatureData
-from lir.data_strategies import PairsTrainTestSplit, PredefinedTrainTestSplit, SourcesTrainTestSplit
 from lir.transform.pairing import InstancePairing
